Move test_offset to directory
